requirements.txt
# LLM APIs
backoff
openai
anthropic
python-dotenv
pymongo
python-ulid
# Viz
matplotlib
pypdf
pymupdf4llm
seaborn
# Common Requirements
numpy
torch
sympy
transformers
datasets
hf-transfer
tiktoken
wandb
tqdm
rich
humanize
dataclasses-json
funcy
black
genson
shutup
python-igraph
coolname
jsonschema
omegaconf
requests
psutil
watchdog
runpod
playwright
